+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [#7821](https://github.com/influxdata/influxdb/issues/7821): Expose some configuration settings via SHOW DIAGNOSTICS
 - [#8025](https://github.com/influxdata/influxdb/issues/8025): Support single and multiline comments in InfluxQL.
 - [#6541](https://github.com/influxdata/influxdb/issues/6541): Support timezone offsets for queries.
+- [#8194](https://github.com/influxdata/influxdb/pull/8194): Add "integral" function to InfluxQL.
 
 ### Bugfixes